✨ Refactor Champions context imports and create ChampionsServerProvider for champion data fetching
